Adani EBook/Ebook/requirements.txt
asgiref==3.6.0
Babel==2.12.1
comtypes==1.1.14
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.7
dj-database-url==2.0.0
Django==4.1.8
django-ajax-datatable==4.4.5
django-environ==0.10.0
django-filter==23.2
django-heroku==0.3.1
django-import-export==3.2.0
django-mathfilters==1.0.0
django-model-utils==1.4.0
django-mssql-backend==2.8.1
django-raw-sugar==0.1.18
django-serverside-datatable==2.1.0
django-tables2==2.5.3
django-tz-detect==0.4.0
djangorestframework==3.14.0
environ==1.0
et-xmlfile==1.1.0
filelock==3.12.2
fuzzywuzzy==0.18.0
MarkupPy==1.14
mssql-django==1.2
numpy==1.24.3
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.0.1
platformdirs==3.10.0
psutil==5.9.5
psycopg2==2.9.6
pyodbc==4.0.39
pypiwin32==223
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2023.3
pywin32==306
PyYAML==6.0
six==1.16.0
sqlparse==0.4.4
tablib==3.4.0
typing_extensions==4.6.3
tzdata==2023.3
virtualenv==20.24.3
whitenoise==6.4.0
xlrd==2.0.1
xlwings==0.30.5
xlwt==1.3.0
